[2023] Day 1: More simplification
Initially, I'd thought to use the non-greedy ".*?" at the end in order
to try to find the last occurrence, but that was failing. Had a shower
thought today that I could just use a the greedy ".*" at the beginning
instead. Theoretically, this _should_ be the same as my initial thought,
but for some reason it works!